Removes Debug module
in favor of Tracing for structured logging and observability purposes. The only shortcoming of this approach is an incompatibility between the command macro and the instrument macro, requiring the implementation being instrumented to be split from the function containing the command macro. It's not ideal, but it suffices. The only alternative to is to accept that spans across wait boundaries will be inaccurate, which is the greater of two evils chosen here

Addresses #27