Fixed race where bot disconnects while queuing (#46)
Due to long queue times as sources buffer from YTDL, the bot may
disconnect if the queue is sitting empty while the next item is still
being enqueued. Since we can't speed up queue times we have two options:
Add the bot back into the channel right before queuing (this is too late)
or add a signal to the disconnect handler which informs if a queue is in
progress. We've gone with the latter in this PR.